+++ CHANGELOG.md
@@ -41,6 +41,7 @@
   * Fixed a wrong Travis-CI configuration,
   * Made some cosmetics changes,
   * Added the sqlite method "query" to micmic mysql library "query" method,
+  * Reverted to sqlite3@5.0.0 as pre-built version does not exist for 5.0.1,
   * ...,