+++ CHANGELOG.md
@@ -42,6 +42,7 @@
   * Made some cosmetics changes,
   * Added the sqlite method "query" to micmic mysql library "query" method,
   * Reverted to sqlite3@5.0.0 as pre-built version does not exist for 5.0.1,
+  * Made some minor changes to the login/logout procedure,
   * ...,